misc: move integration tabs options to a const file
