+++ src/components/PaymentProviderChip.tsx
@@ -3,6 +3,7 @@ import { FC } from 'react'
 import { ProviderTypeEnum } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
 import Adyen from '~/public/images/adyen.svg'
+import Cashfree from '~/public/images/cashfree.svg'
 import Gocardless from '~/public/images/gocardless.svg'
 import Stripe from '~/public/images/stripe.svg'
 import { tw } from '~/styles/utils'
@@ -26,6 +27,10 @@ const providers: Record<ProviderTypeEnum, { icon: JSX.Element; label: string }> 
   [ProviderTypeEnum.Gocardless]: {
     icon: <Gocardless />,
     label: 'text_634ea0ecc6147de10ddb6625',
+  },
+  [ProviderTypeEnum.Cashfree]: {
+    icon: <Cashfree />,
+    label: 'text_17367626793434wkg1rk0114',
   },
 }
