+++ src/components/plans/details/PlanSubscriptionListItem.tsx
@@ -5,10 +5,10 @@ import styled, { css } from 'styled-components'
 
 import { computeCustomerInitials } from '~/components/customers/utils'
 import { Avatar, Skeleton, Typography } from '~/components/designSystem'
+import { CustomerSubscriptionDetailsTabsOptionsEnum } from '~/core/constants/tabsOptions'
 import { PLAN_SUBSCRIPTION_DETAILS_ROUTE } from '~/core/router'
 import { PlanSubscriptionListItemForSubscriptionListFragment } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
-import { CustomerSubscriptionDetailsTabsOptionsEnum } from '~/pages/SubscriptionDetails'
 import { ListItemLink, NAV_HEIGHT, theme } from '~/styles'
 
 gql`
