+++ src/components/settings/integrations/AddGocardlessDialog.tsx
@@ -11,6 +11,7 @@ import { Button, Dialog, DialogRef } from '~/components/designSystem'
 import { TextInputField } from '~/components/form'
 import { addToast, envGlobalVar } from '~/core/apolloClient'
 import { buildGocardlessAuthUrl } from '~/core/constants/externalUrls'
+import { IntegrationsTabsOptionsEnum } from '~/core/constants/tabsOptions'
 import { GOCARDLESS_INTEGRATION_DETAILS_ROUTE } from '~/core/router'
 import {
   AddGocardlessPaymentProviderInput,
@@ -21,7 +22,6 @@ import {
   useUpdateGocardlessApiKeyMutation,
 } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
-import { IntegrationsTabsOptionsEnum } from '~/pages/settings/Integrations'
 import { theme } from '~/styles'
 
 import { DeleteGocardlessIntegrationDialogRef } from './DeleteGocardlessIntegrationDialog'
