+++ src/components/settings/integrations/AddSalesforceDialog.tsx
@@ -8,6 +8,7 @@ import { object, string } from 'yup'
 import { Alert, Button, Dialog, DialogRef } from '~/components/designSystem'
 import { TextInputField } from '~/components/form'
 import { addToast, hasDefinedGQLError } from '~/core/apolloClient'
+import { IntegrationsTabsOptionsEnum } from '~/core/constants/tabsOptions'
 import { SALESFORCE_INTEGRATION_DETAILS_ROUTE } from '~/core/router'
 import {
   CreateSalesforceIntegrationInput,
@@ -17,7 +18,6 @@ import {
   useUpdateSalesforceIntegrationMutation,
 } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
-import { IntegrationsTabsOptionsEnum } from '~/pages/settings/Integrations'
 import { tw } from '~/styles/utils'
 
 import { DeleteSalesforceIntegrationDialogRef } from './DeleteSalesforceIntegrationDialog'
