+++ src/core/constants/tabsOptions.ts
@@ -1,0 +1,19 @@
+export enum CustomerInvoiceDetailsTabsOptionsEnum {
+  overview = 'overview',
+  creditNotes = 'credit-notes',
+}
+
+export enum CustomerSubscriptionDetailsTabsOptionsEnum {
+  overview = 'overview',
+  usage = 'usage',
+}
+
+export enum IntegrationsTabsOptionsEnum {
+  Lago = 'lago',
+  Community = 'community',
+}
+
+export enum PlanDetailsTabsOptionsEnum {
+  overview = 'overview',
+  subscriptions = 'subscriptions',
+}
