+++ src/pages/settings/HubspotIntegrations.tsx
@@ -21,6 +21,7 @@ import {
   DeleteHubspotIntegrationDialog,
   DeleteHubspotIntegrationDialogRef,
 } from '~/components/settings/integrations/DeleteHubspotIntegrationDialog'
+import { IntegrationsTabsOptionsEnum } from '~/core/constants/tabsOptions'
 import { HUBSPOT_INTEGRATION_DETAILS_ROUTE, INTEGRATIONS_ROUTE } from '~/core/router'
 import {
   HubspotForCreateDialogFragmentDoc,
@@ -31,8 +32,6 @@ import {
 import { useInternationalization } from '~/hooks/core/useInternationalization'
 import Hubspot from '~/public/images/hubspot.svg'
 import { ListItemLink, MenuPopper, PageHeader, PopperOpener } from '~/styles'
-
-import { IntegrationsTabsOptionsEnum } from './Integrations'
 
 gql`
   fragment HubspotIntegrations on HubspotIntegration {
