+++ translations/base.json
@@ -2790,5 +2790,6 @@
   "text_1733303404276jppxvximavl": "Built by Lago",
   "text_173330340427732b341qnuny": "Built by community",
   "text_1733303404277q80b216p5zr": "This integration is not officially developed by Lago. Please note that the Lago team will not be able to provide support if any issues arise.",
-  "text_1733303818769298k0fvsgcz": "Type a redirect URL"
+  "text_1733303818769298k0fvsgcz": "Type a redirect URL",
+  "text_17367626793434wkg1rk0114": "Cashfree"
 }
